deckforge/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-pptx==0.6.23
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
google-generativeai==0.5.0
openai>=1.0.0
email-validator>=2.0.0
motor==3.4.0
pymongo==4.8.0
PyPDF2>=3.0.0
python-docx>=0.8.11
certifi>=2024.0.0